solution/requirements.txt
torch
torchvision
numpy
pillow
matplotlib
json
opencv-python
tqdm